refactor: move gradient colors helper to ui/theme
