+++ src/ui/theme/getGradientColors.ts
